Add List check + combine modules
Add a check for list endpoint
Combine the userModules and defaultModules so that only activate is needed to get all endpoints
