+++ src/combine-modules.js
@@ -1,0 +1,43 @@
+const defaultModules = {
+  account: {
+    activate: true,
+    endpoints: {
+      tvs: ['accountFavoriteTv', 'accountRatedTv', 'accountTvWatchlist'],
+      movies: ['accountFavoriteMovies', 'accountRatedMovies', 'accountMovieWatchlist'],
+      list: 'accountLists',
+    },
+  },
+  misc: {
+    activate: false,
+    endpoints: [
+      ['miscUpcomingMovies'],
+      ['miscNowPlayingMovies'],
+      ['miscPopularMovies', 2],
+      ['miscTopRatedMovies', 2],
+      ['miscTopRatedTvs', 1],
+      ['miscPopularTvs', 1],
+    ],
+  },
+  tv: {
+    activate: false,
+    endpoints: [['tvAiringToday'], ['tvOnTheAir', 2]],
+  },
+}
+
+const combineModules = userModules =>
+  Object.assign(defaultModules, {
+    account: {
+      ...defaultModules.account,
+      ...userModules.account,
+    },
+    misc: {
+      ...defaultModules.misc,
+      ...userModules.misc,
+    },
+    tv: {
+      ...defaultModules.tv,
+      ...userModules.tv,
+    },
+  })
+
+exports.combineModules = combineModules
